リファクタリング: find関数を切り出し
